feat(node): systematic type safety and test enrichment across 56 modules
Concurrent agent sweep applying targeted improvements across the entire
franken-node codebase. All changes are within existing modules — no new
production logic, only safety hardening and test additions.

Categories of changes across 56 Rust source files:

Connector subsystem (18 modules):
- admission_budget, bocpd, canonical_serializer, capability_artifact,
  crdt, device_profile, durable_claim_gate, ecosystem_compliance,
  incident_bundle_retention, lease_conflict, obligation_tracker,
  policy_checkpoint, quarantine_store, retention_policy, trust_fabric,
  universal_verifier_sdk, vef_execution_receipt, verifier_sdk

Control plane (8 modules):
- audience_token, cancellation_injection, control_lane_mapping,
  dpor_exploration, fork_detection, marker_stream, root_pointer,
  transition_abort

Security (5 modules):
- bpet/economic_integration, constant_time, decision_receipt,
  revocation_freshness, vef_degraded_mode

Supply chain (5 modules):
- extension_registry, migration_kit, revocation_integration,
  revocation_registry, trust_card

Tools/testing (7 modules):
- benchmark_methodology, benchmark_suite, migration_validation_cohorts,
  redteam_evaluations, repro_bundle_export, transparent_reports,
  verifier_benchmark_releases

Other domains (8 modules):
- policy, registry, remote, runtime, observability, testing,
  vef, verifier_economy

+++ crates/franken-node/src/connector/crdt.rs
@@ -251,8 +251,8 @@ impl PnCounter {
         self.negative.increment(replica_id, amount);
     }
 
-    pub fn value(&self) -> i64 {
-        self.positive.value() as i64 - self.negative.value() as i64
+    pub fn value(&self) -> i128 {
+        self.positive.value() as i128 - self.negative.value() as i128
     }
 
     pub fn merge(&self, other: &PnCounter) -> Result<PnCounter, CrdtError> {
